moved logic to helper class
